requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
jinja2==3.1.4
python-multipart==0.0.12
sqlalchemy==2.0.36
aiosqlite==0.20.0
asyncpg==0.30.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
python-jose[cryptography]==3.3.0
anthropic==0.40.0
stripe==11.3.0
python-dotenv==1.0.1
httpx==0.28.1
email-validator==2.2.0
slowapi==0.1.9
reportlab==4.2.5
